docs/api: Configure tools for Open API spec
Create a Makefile to install and execute all the different tools used
for being able to work on an Open API specification split in several
files rather than in one single file for avoiding conflicts while
working several people on it at the same time.

The commit also includes configuration files required for installing and
running the tools.