payment: Simplify impl reducing flexibility
Simplify the implementation reducing features regarding the number of
fields which can be selected or not and the fields which the collections
of payment can be sorted.

The amounts fields of the payment has also be changed to be basic float
types for easing the conversion and because big numbers aren't needed
due to no need to make numeric operations with those.

This mostly done because this is an example of a project so I'm
simplifying for the sake of finishing the implementation part and have
it ready for experimentations in different areas.