Makefile: Remove duplicated target & arrangements
Remove the duplicated target that the last rebase from master has
brought and make some arrangements regarding the tools that the CI must
install for being able to run the 'ci' target, which is the last change
that the CI configuration suffered in master and it has been brought to
this point due to the rebase.

On the other hand the 'ci' target needed to be tweaked because the
previous coverage mode doesn't work with the `-race` go test flag.