docs/api: Replace unsigned num formats
The OpenAPI spec v3 doesn't support unsigned numbers in the format field
of a number type, so they have been replaced by signed ones.

I was wrong using them because in the past I used go-swagger, which
accepts to use the unsigned format in order of having an exact mapping
with the Go types, however I don't plan to use go-swagger in this
project because it doesn't support OpenAPI v3, just only v2.